Make "all" archive easier to test
